requirements.txt
streamlit>=1.34
tensorflow-cpu>=2.16
numpy>=1.26
pillow>=10.3
requests>=2.32
